Split localization module into multiple files
